codex: move synthetic-response feature to dedicated openai_codex_https endpoint
The first cut wired the JWKS + agent-task-register stubs and the
CODEX_ACCESS_TOKEN env push-down through the openai_codex_oauth
*credential* plugin, so any endpoint binding that credential —
including api.openai.com — would inherit the chatgpt.com-only
behavior. Move the lot to a dedicated endpoint plugin instead.

Operators now wire codex with:

    credential "openai_codex_oauth" "codex" {}

    endpoint "openai_codex_https" "codex" {
      hosts      = ["chatgpt.com"]
      credential = codex
    }

* New endpoint plugin openai_codex_https owns the JWT/JWKS keypair,
  EnvVars (CODEX_ACCESS_TOKEN + CODEX_AGENT_IDENTITY +
  CODEX_AGENT_IDENTITY_AUTHAPI_BASE_URL), and the synthetic-response
  paths. Default https plugin is back to a generic, codex-free shape.
* HTTPCredentialResponder interface dropped; replaced with
  HTTPSyntheticResponder on the endpoint runtime. mitmHTTPS
  dispatches the synth hook off ep.Plugin.Runtime before credential
  injection.
* Env push-down now iterates KindEndpoint plugins as well as
  KindCredential, so endpoint-shaped envs (routing controls, base
  URL overrides) emit the same way credential placeholders do.
* Starter policy (login.go) splits the old combined openai endpoint
  into an https binding for api.openai.com plus an
  openai_codex_https binding for chatgpt.com.
