+++ config/plugins/credentials/openai_codex.go
@@ -4,11 +4,9 @@ package credentials
 // api.openai.com + chatgpt.com both accept Authorization: Bearer.
 
 import (
-	"bytes"
 	"context"
 	"encoding/base64"
 	"encoding/json"
-	"io"
 	"net/http"
 	"strings"
 
@@ -66,96 +64,21 @@ func chatgptAccountID(jwt string) string {
 	return claims.Auth.ChatGPTAccountID
 }
 
-// EnvVars pushes a synthetic CODEX_ACCESS_TOKEN (Agent Identity JWT
-// signed by clawpatrol) plus a redirect for the agent-identity API
-// base URL. Both bypass the OPENAI_API_KEY codepath that would
-// otherwise route codex to api.openai.com — the synthetic JWT puts
-// codex into AgentIdentity mode and points it at chatgpt.com, where
-// the existing chatgpt.com binding swaps Authorization +
-// chatgpt-account-id for the user's real subscription bearer.
+// EnvVars intentionally returns nothing.
 //
-// The JWT validates against a JWKS clawpatrol serves locally (see
-// RespondHTTP below) since clawpatrol owns chatgpt.com's TLS via
-// MITM. Task registration is similarly stubbed locally — pushing
-// CODEX_AGENT_IDENTITY_AUTHAPI_BASE_URL to chatgpt.com keeps that
-// call on a host we already terminate, instead of leaking to
-// auth.openai.com (codex's default).
-func (*OpenAICodexOAuth) EnvVars() []config.EnvVar {
-	jwt, err := MintCodexAccessToken()
-	if err != nil {
-		// Fall back silently — without the JWT codex goes through its
-		// real auth.json and clawpatrol's MITM still works for users
-		// who already ran `codex login`. The error surfaces in
-		// `clawpatrol env`'s stderr via the caller's logging.
-		return nil
-	}
-	return []config.EnvVar{
-		{
-			Name:        "CODEX_ACCESS_TOKEN",
-			Value:       jwt,
-			Description: "synthetic Agent Identity JWT — routes codex ≥ 0.129 to chatgpt.com",
-		},
-		{
-			// Codex 0.128 and earlier read the JWT from
-			// CODEX_AGENT_IDENTITY; the rename to CODEX_ACCESS_TOKEN
-			// landed post-0.128 (codex commit 0d418f478d). Set both so
-			// clawpatrol works across versions; whichever the
-			// installed codex looks for wins.
-			Name:        "CODEX_AGENT_IDENTITY",
-			Value:       jwt,
-			Description: "synthetic Agent Identity JWT — routes codex ≤ 0.128 to chatgpt.com",
-		},
-		{
-			Name:        "CODEX_AGENT_IDENTITY_AUTHAPI_BASE_URL",
-			Value:       "https://chatgpt.com/backend-api/wham",
-			Description: "keeps agent-task registration on a host clawpatrol MITMs",
-		},
-	}
-}
-
-// RespondHTTP intercepts the two chatgpt.com paths codex hits during
-// Agent Identity init: the JWKS that validates the synthetic JWT and
-// the agent-task registration POST that returns a task_id. Both are
-// served from clawpatrol-controlled state — neither reaches the real
-// chatgpt.com.
-func (a *OpenAICodexOAuth) RespondHTTP(_ context.Context, req *http.Request, _ runtime.Secret) (*http.Response, bool, error) {
-	if !strings.EqualFold(req.URL.Host, "chatgpt.com") && !strings.EqualFold(req.Host, "chatgpt.com") {
-		return nil, false, nil
-	}
-	switch {
-	case req.Method == http.MethodGet && req.URL.Path == "/backend-api/wham/agent-identities/jwks":
-		body, err := CodexJWKSResponse()
-		if err != nil {
-			return nil, false, err
-		}
-		return jsonResp(req, http.StatusOK, body), true, nil
-	case req.Method == http.MethodPost && strings.HasPrefix(req.URL.Path, "/backend-api/wham/v1/agent/") &&
-		strings.HasSuffix(req.URL.Path, "/task/register"):
-		return jsonResp(req, http.StatusOK, []byte(`{"task_id":"clawpatrol-task"}`)), true, nil
-	}
-	return nil, false, nil
-}
-
-// jsonResp builds an http.Response the gateway can write back to the
-// agent. We bypass http.Transport entirely for synthetic responses
-// — the response is constructed in-memory and flushed by the caller
-// via http.Response.Write.
-func jsonResp(req *http.Request, status int, body []byte) *http.Response {
-	resp := &http.Response{
-		Status:        http.StatusText(status),
-		StatusCode:    status,
-		Proto:         "HTTP/1.1",
-		ProtoMajor:    1,
-		ProtoMinor:    1,
-		Header:        make(http.Header),
-		Body:          io.NopCloser(bytes.NewReader(body)),
-		ContentLength: int64(len(body)),
-		Request:       req,
-	}
-	resp.Header.Set("Content-Type", "application/json")
-	resp.Header.Set("Cache-Control", "no-store")
-	return resp
-}
+// Codex env-var push-down lives on the openai_codex_https endpoint
+// plugin (config/plugins/endpoints/openai_codex_https.go). Pushing it
+// from the credential would attach the synthetic JWT to every
+// endpoint that binds this credential — including api.openai.com,
+// where it has no business going. Operators wire codex via:
+//
+//	credential "openai_codex_oauth" "codex" {}
+//
+//	endpoint "openai_codex_https" "codex" {
+//	  hosts      = ["chatgpt.com"]
+//	  credential = codex
+//	}
+func (*OpenAICodexOAuth) EnvVars() []config.EnvVar { return nil }
 
 func (a *OpenAICodexOAuth) OAuthFlow() *config.OAuthIntegration {
 	return &config.OAuthIntegration{
@@ -184,7 +107,6 @@ func (a *OpenAICodexOAuth) OAuthFlow() *config.OAuthIntegration {
 
 func init() {
 	var _ runtime.HTTPCredentialRuntime = (*OpenAICodexOAuth)(nil)
-	var _ runtime.HTTPCredentialResponder = (*OpenAICodexOAuth)(nil)
 	config.Register(&config.Plugin{
 		Kind:    config.KindCredential,
 		Type:    "openai_codex_oauth",
