+++ config/plugins/endpoints/openai_codex_https_test.go
@@ -1,4 +1,4 @@
-package credentials
+package endpoints
 
 import (
 	"context"
@@ -17,8 +17,6 @@ import (
 	"strings"
 	"sync"
 	"testing"
-
-	"github.com/denoland/clawpatrol/config/runtime"
 )
 
 // TestCodexJWTRoundTrip mints a JWT and verifies its RS256 signature
@@ -38,7 +36,7 @@ func TestCodexJWTRoundTrip(t *testing.T) {
 	codexKeys = nil
 	codexKeysErr = nil
 
-	jwt, err := MintCodexAccessToken()
+	jwt, err := mintCodexAccessToken()
 	if err != nil {
 		t.Fatalf("mint: %v", err)
 	}
@@ -49,18 +47,16 @@ func TestCodexJWTRoundTrip(t *testing.T) {
 		t.Fatalf("keys file missing: %v", err)
 	}
 
-	jwksJSON, err := CodexJWKSResponse()
+	jwksJSON, err := codexJWKSResponse()
 	if err != nil {
 		t.Fatalf("jwks: %v", err)
 	}
 
-	// Parse JWT into segments
 	parts := strings.Split(jwt, ".")
 	if len(parts) != 3 {
 		t.Fatalf("jwt has %d parts, want 3", len(parts))
 	}
 
-	// Verify the kid in the JWT header is in the JWKS we serve.
 	headerJSON, err := base64.RawURLEncoding.DecodeString(parts[0])
 	if err != nil {
 		t.Fatalf("decode header: %v", err)
@@ -101,7 +97,6 @@ func TestCodexJWTRoundTrip(t *testing.T) {
 		t.Fatalf("kid %q not in JWKS", hdr.Kid)
 	}
 
-	// Verify signature
 	signingInput := parts[0] + "." + parts[1]
 	sig, err := base64.RawURLEncoding.DecodeString(parts[2])
 	if err != nil {
@@ -112,7 +107,6 @@ func TestCodexJWTRoundTrip(t *testing.T) {
 		t.Fatalf("verify: %v", err)
 	}
 
-	// Verify required claims
 	claimsJSON, err := base64.RawURLEncoding.DecodeString(parts[1])
 	if err != nil {
 		t.Fatalf("decode claims: %v", err)
@@ -146,18 +140,49 @@ func TestCodexJWTRoundTrip(t *testing.T) {
 	}
 }
 
+// TestCodexEndpointEnvVars verifies the endpoint plugin emits the
+// CODEX_ACCESS_TOKEN / CODEX_AGENT_IDENTITY env vars (both for cross-
+// version codex compat) plus the auth-API base URL override.
+func TestCodexEndpointEnvVars(t *testing.T) {
+	t.Setenv("CLAWPATROL_DIR", t.TempDir())
+	codexKeysOnce = sync.Once{}
+	codexKeys = nil
+	codexKeysErr = nil
+
+	got := (&OpenAICodexHTTPSEndpoint{}).EnvVars()
+	want := map[string]bool{
+		"CODEX_ACCESS_TOKEN":                    false,
+		"CODEX_AGENT_IDENTITY":                  false,
+		"CODEX_AGENT_IDENTITY_AUTHAPI_BASE_URL": false,
+	}
+	for _, ev := range got {
+		if _, ok := want[ev.Name]; !ok {
+			t.Errorf("unexpected env var: %s", ev.Name)
+			continue
+		}
+		want[ev.Name] = true
+		if ev.Value == "" {
+			t.Errorf("%s has empty value", ev.Name)
+		}
+	}
+	for n, ok := range want {
+		if !ok {
+			t.Errorf("missing env var: %s", n)
+		}
+	}
+}
+
 // TestCodexRespondHTTP exercises the synthetic-response paths the
-// codex plugin handles: JWKS fetch + agent task registration. Both
-// must return 200 with parseable JSON; non-matching paths must fall
-// through. We also confirm a request on a non-chatgpt host (the
-// api.openai.com side of the binding) is left for normal forwarding.
+// endpoint runtime handles: JWKS fetch + agent task registration.
+// Both must return 200 with parseable JSON; non-matching paths must
+// fall through.
 func TestCodexRespondHTTP(t *testing.T) {
 	t.Setenv("CLAWPATROL_DIR", t.TempDir())
 	codexKeysOnce = sync.Once{}
 	codexKeys = nil
 	codexKeysErr = nil
 
-	plugin := &OpenAICodexOAuth{}
+	rt := OpenAICodexHTTPSEndpointRuntime{}
 	cases := []struct {
 		name    string
 		method  string
@@ -168,7 +193,7 @@ func TestCodexRespondHTTP(t *testing.T) {
 		{"jwks", "GET", "https://chatgpt.com/backend-api/wham/agent-identities/jwks", 200, true},
 		{"register", "POST", "https://chatgpt.com/backend-api/wham/v1/agent/clawpatrol-codex/task/register", 200, true},
 		{"forward responses", "POST", "https://chatgpt.com/backend-api/codex/responses", 0, false},
-		{"openai api", "POST", "https://api.openai.com/v1/responses", 0, false},
+		{"unrelated path", "GET", "https://chatgpt.com/", 0, false},
 	}
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
@@ -177,7 +202,7 @@ func TestCodexRespondHTTP(t *testing.T) {
 				t.Fatalf("url: %v", err)
 			}
 			req := &http.Request{Method: c.method, URL: u, Host: u.Host, Header: http.Header{}}
-			resp, handled, err := plugin.RespondHTTP(context.Background(), req, runtime.Secret{})
+			resp, handled, err := rt.RespondHTTP(context.Background(), req)
 			if err != nil {
 				t.Fatalf("respond: %v", err)
 			}
@@ -194,75 +219,27 @@ func TestCodexRespondHTTP(t *testing.T) {
 			if err != nil {
 				t.Fatalf("body: %v", err)
 			}
-			var any map[string]any
-			if err := json.Unmarshal(body, &any); err != nil {
+			var anyMap map[string]any
+			if err := json.Unmarshal(body, &anyMap); err != nil {
 				t.Fatalf("parse body %q: %v", body, err)
 			}
 		})
 	}
 }
 
-// TestCodexInjectOverridesAgentAssertion confirms the auth swap path:
-// codex in Agent Identity mode sends `Authorization: AgentAssertion …`
-// signed with the per-task ed25519 key from the synthetic JWT we
-// minted. The MITM must overwrite that with `Bearer <real subscription
-// token>` plus the real account-id from the bearer's JWT claims, so
-// chatgpt.com sees a request indistinguishable from native ChatGPT
-// auth. Without this swap, the upstream rejects with 403 "Unknown
-// agent runtime for AgentAssertion" (verified live against funk).
-func TestCodexInjectOverridesAgentAssertion(t *testing.T) {
-	plugin := &OpenAICodexOAuth{}
-
-	// A real-shaped chatgpt access_token JWT carrying chatgpt_account_id
-	// in the auth namespace claim — same form as
-	// ~/.codex/auth.json's tokens.access_token.
-	headerJSON, _ := json.Marshal(map[string]string{"alg": "RS256", "typ": "JWT"})
-	claimsJSON, _ := json.Marshal(map[string]any{
-		"https://api.openai.com/auth": map[string]string{
-			"chatgpt_account_id": "real-acct-7b415a8c",
-		},
-	})
-	bearer := base64.RawURLEncoding.EncodeToString(headerJSON) + "." +
-		base64.RawURLEncoding.EncodeToString(claimsJSON) + "." +
-		base64.RawURLEncoding.EncodeToString([]byte("sig"))
-
-	req, err := http.NewRequest("POST", "https://chatgpt.com/backend-api/codex/responses", strings.NewReader("{}"))
-	if err != nil {
-		t.Fatalf("new req: %v", err)
-	}
-	// What codex (Agent Identity mode) would have sent
-	req.Header.Set("Authorization", "AgentAssertion eyJhZ2VudF9ydW50aW1lX2lkIjoiY2xhd3BhdHJvbC1jb2RleCIsInNpZ25hdHVyZSI6Im9wYXF1ZSJ9")
-	req.Header.Set("Chatgpt-Account-Id", "clawpatrol")
-
-	if err := plugin.InjectHTTP(req.Context(), req, runtime.Secret{Bytes: []byte(bearer)}); err != nil {
-		t.Fatalf("inject: %v", err)
-	}
-	if got := req.Header.Get("Authorization"); got != "Bearer "+bearer {
-		t.Errorf("Authorization not swapped to bearer: got %q", got)
-	}
-	if got := req.Header.Get("Chatgpt-Account-Id"); got != "real-acct-7b415a8c" {
-		t.Errorf("Chatgpt-Account-Id not overwritten with bearer claim: got %q", got)
-	}
-}
-
 // TestCodexSynthRoundTripOverHTTP wires the synthetic responder
 // through net/http/httptest so we can issue a real HTTP GET and
-// confirm the bytes the agent sees are what we serve. This is the
-// closest thing to an integration test without standing up the full
-// MITM gateway.
+// confirm the bytes the agent sees are what we serve. Closest thing
+// to an integration test without standing up the full MITM gateway.
 func TestCodexSynthRoundTripOverHTTP(t *testing.T) {
 	t.Setenv("CLAWPATROL_DIR", t.TempDir())
 	codexKeysOnce = sync.Once{}
 	codexKeys = nil
 	codexKeysErr = nil
 
-	plugin := &OpenAICodexOAuth{}
+	rt := OpenAICodexHTTPSEndpointRuntime{}
 	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// httptest serves under 127.0.0.1; rewrite Host so the plugin's
-		// chatgpt.com guard fires.
-		r.Host = "chatgpt.com"
-		r.URL.Host = "chatgpt.com"
-		resp, handled, err := plugin.RespondHTTP(r.Context(), r, runtime.Secret{})
+		resp, handled, err := rt.RespondHTTP(r.Context(), r)
 		if err != nil {
 			http.Error(w, err.Error(), 500)
 			return
