+++ config/runtime/runtime.go
@@ -30,20 +30,25 @@ type HTTPCredentialRuntime interface {
 	InjectHTTP(ctx context.Context, req *http.Request, sec Secret) error
 }
 
-// HTTPCredentialResponder is the optional contract a credential plugin
-// implements when it needs to short-circuit certain matched requests
-// and return a synthetic response without forwarding upstream. Codex
-// uses this to serve a clawpatrol-controlled JWKS at chatgpt.com's
-// agent-identity URL (so a JWT we minted ourselves validates) and to
-// stub out the agent-task registration POST.
+// HTTPSyntheticResponder is the optional contract an endpoint
+// plugin's runtime implements when it needs to short-circuit certain
+// matched requests and return a synthetic response without forwarding
+// upstream. The openai_codex_https endpoint uses it to serve a
+// clawpatrol-controlled JWKS at chatgpt.com's agent-identity URL (so
+// a JWT we minted ourselves validates) and to stub out the agent-task
+// registration POST.
 //
-// RespondHTTP is called after credential injection but before the
-// upstream forward. Returning (resp, true, nil) writes resp to the
-// agent and skips forwarding; returning (_, false, nil) falls through
-// to the normal proxy path. Errors are logged and the request still
+// RespondHTTP is called by mitmHTTPS before credential injection.
+// Returning (resp, true, nil) writes resp to the agent and skips
+// forwarding; returning (_, false, nil) falls through to the normal
+// inject + proxy path. Errors are logged and the request still
 // forwards verbatim.
-type HTTPCredentialResponder interface {
-	RespondHTTP(ctx context.Context, req *http.Request, sec Secret) (*http.Response, bool, error)
+//
+// The hook lives on the endpoint plugin (not the credential) so the
+// behavior is bound to the protocol surface, not to whichever bearer
+// happens to be configured for it.
+type HTTPSyntheticResponder interface {
+	RespondHTTP(ctx context.Context, req *http.Request) (*http.Response, bool, error)
 }
 
 // PostgresCredentialRuntime swaps the agent's StartupMessage password
