+++ main.go
@@ -1264,6 +1264,28 @@ func (g *Gateway) mitmHTTPS(c net.Conn, host string, ep *config.CompiledEndpoint
 			}
 		}
 
+		// Endpoint-level synthetic-response hook. The endpoint
+		// plugin's runtime can short-circuit specific paths and
+		// return a clawpatrol-generated response without forwarding
+		// upstream — used by openai_codex_https to serve the JWKS +
+		// agent-task-register stubs that anchor codex's Agent
+		// Identity flow on hosts we MITM. Endpoints without a
+		// responder (the default https plugin) fall through.
+		if responder, ok := ep.Plugin.Runtime.(runtime.HTTPSyntheticResponder); ok {
+			if r, handled, err := responder.RespondHTTP(req.Context(), req); err != nil {
+				log.Printf("respond %s: %v", ep.Name, err)
+			} else if handled {
+				ev.Status = r.StatusCode
+				ev.Action = "synth"
+				if err := r.Write(tc); err != nil {
+					log.Printf("synth write %s %s: %v", host, req.URL.Path, err)
+				}
+				ev.Ms = time.Since(start).Milliseconds()
+				g.emitEnd(ev)
+				continue
+			}
+		}
+
 		// Credential injection. Pick the credential entry that
 		// applies to this request (singular binding short-circuits;
 		// multi-credential dispatch asks the endpoint plugin's
@@ -1273,51 +1295,22 @@ func (g *Gateway) mitmHTTPS(c net.Conn, host string, ep *config.CompiledEndpoint
 		// to stamp onto the request. Schema-only credential types
 		// (slack / telegram / gemini / etc.) leave Runtime nil; we
 		// pass through verbatim and rely on policy alone.
-		var synthResp *http.Response
 		if cc := runtime.ResolveCredential(ep, mreq); cc != nil {
 			// Plugin.Runtime is a typed-nil sentinel used only for
 			// interface-compliance assertions; the actual decoded HCL
 			// values (BearerToken.IdempotencyKey, PostgresCredential.User,
 			// etc.) live on Body. Invoke methods through Body so the
 			// receiver is the real instance.
-			var sec runtime.Secret
-			var secErr error
-			if _, ok := cc.Credential.Body.(runtime.HTTPCredentialRuntime); ok {
-				sec, secErr = g.secrets.Get(cc.Credential.Symbol.Name, profile)
-			}
-			// Synthetic-response hook (e.g. codex's JWKS / task-register
-			// stubs). Runs before the upstream forward — credential
-			// plugins use this to terminate well-known internal paths
-			// without leaking to the network. Errors fall through to
-			// normal proxying.
-			if responder, ok := cc.Credential.Body.(runtime.HTTPCredentialResponder); ok {
-				if r, handled, err := responder.RespondHTTP(req.Context(), req, sec); err != nil {
-					log.Printf("respond %s: %v", cc.Credential.Symbol.Name, err)
-				} else if handled {
-					synthResp = r
+			if injector, ok := cc.Credential.Body.(runtime.HTTPCredentialRuntime); ok {
+				sec, err := g.secrets.Get(cc.Credential.Symbol.Name, profile)
+				if err != nil {
+					log.Printf("secret %s/%s: %v — forwarding without injection", cc.Credential.Symbol.Name, profile, err)
+				} else if len(sec.Bytes) == 0 {
+					log.Printf("secret %s/%s: not configured (set CLAWPATROL_SECRET_%s)", cc.Credential.Symbol.Name, profile, secretEnvName(cc.Credential.Symbol.Name))
+				} else if err := injector.InjectHTTP(req.Context(), req, sec); err != nil {
+					log.Printf("inject %s: %v", cc.Credential.Symbol.Name, err)
 				}
 			}
-			if synthResp == nil {
-				if injector, ok := cc.Credential.Body.(runtime.HTTPCredentialRuntime); ok {
-					if secErr != nil {
-						log.Printf("secret %s/%s: %v — forwarding without injection", cc.Credential.Symbol.Name, profile, secErr)
-					} else if len(sec.Bytes) == 0 {
-						log.Printf("secret %s/%s: not configured (set CLAWPATROL_SECRET_%s)", cc.Credential.Symbol.Name, profile, secretEnvName(cc.Credential.Symbol.Name))
-					} else if err := injector.InjectHTTP(req.Context(), req, sec); err != nil {
-						log.Printf("inject %s: %v", cc.Credential.Symbol.Name, err)
-					}
-				}
-			}
-		}
-		if synthResp != nil {
-			ev.Status = synthResp.StatusCode
-			ev.Action = "synth"
-			if err := synthResp.Write(tc); err != nil {
-				log.Printf("synth write %s %s: %v", host, req.URL.Path, err)
-			}
-			ev.Ms = time.Since(start).Milliseconds()
-			g.emitEnd(ev)
-			continue
 		}
 
 		// WebSocket upgrade. http.Transport.RoundTrip mangles the
